restructure methods + wrappers
